People: Move EditUserForm to separate file
